Move `heck` and `sane` checks into `tests/utils.py`
